requirements.txt
aiogram==3.7.0
aiofiles==23.2.1
sqlalchemy==2.0.23
alembic==1.12.1
apscheduler==3.10.4
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.19.0 